\name{ag.ds.heatmapplot}
\alias{ag.ds.heatmapplot}
\title{Generates a heatmap plot for merged datasets}
\usage{
  ag.ds.heatmapplot(opals, xvect, yvect)
}
\arguments{
  \item{opals}{a list of opal object(s) obtained after
  login in to opal servers; these objects hold also the
  data assign to R, as \code{dataframe}, from opal
  datasources.}

  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a heatmap plot
}
\description{
  Generates a heatmap plot for merged datasets
}
\author{
  Isaeva, J.
}

