\name{ag.ds.histogram}
\alias{ag.ds.histogram}
\title{Plots a histogram}
\usage{
  ag.ds.histogram(opals, a)
}
\arguments{
  \item{opals}{a character strings that represent the URL
  of the servers where the study datasets are stored}

  \item{a}{vector of values for which the histogram is
  desired.}
}
\value{
  a histogram.
}
\description{
  Computes and plots histogram of the given data values.
}
\author{
  Burton, P.
}

