\name{ds.histogram}
\alias{ds.histogram}
\title{Plots a histogram}
\usage{
  ds.histogram(opals, a)
}
\arguments{
  \item{a}{vector of values for which the histogram is
  desired.}
}
\value{
  a histogram.
}
\description{
  Computes and plots histogram of the given data values.
}
\author{
  Burton, P.
}

